move remaining plugins and cleanup
