+++ bokeh/testing/__init__.py
@@ -5,7 +5,7 @@
 #
 # The full license is in the file LICENSE.txt, distributed with this software.
 #-----------------------------------------------------------------------------
-''' Provide suport modules for testing Bokeh itself.
+''' Provide support modules for testing Bokeh itself.
 
 '''
 
@@ -26,6 +26,10 @@ log = logging.getLogger(__name__)
 # External imports
 
 # Bokeh imports
+
+#-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
 
 #-----------------------------------------------------------------------------
 # General API
