+++ bokeh/testing/examples.py
@@ -1,17 +1,47 @@
-from __future__ import absolute_import, print_function
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2017, Anaconda, Inc. All rights reserved.
+#
+# Powered by the Bokeh Development Team.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+''' Provide suport modules for testing Bokeh itself.
 
+'''
+
+#-----------------------------------------------------------------------------
+# Boilerplate
+#-----------------------------------------------------------------------------
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+import logging
+log = logging.getLogger(__name__)
+
+#-----------------------------------------------------------------------------
+# Imports
+#-----------------------------------------------------------------------------
+
+# Standard library imports
 import os
-
+from os.path import join, exists, dirname, basename, relpath, splitext, isfile, isdir
 import yaml
+
+# External imports
 import requests
 
-from os.path import join, exists, dirname, basename, relpath, splitext, isfile, isdir
-
-from ..plugins.upload_to_s3 import S3_URL, upload_file_to_s3
-
+# Bokeh imports
 from bokeh.testing.git import __version__
+from bokeh.testing.s3 import S3_URL, upload_file_to_s3
 from bokeh.testing.travis import JOB_ID
 from bokeh.util.terminal import trace, green
+
+#-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# General API
+#-----------------------------------------------------------------------------
 
 class Flags(object):
     js       = 1 << 0
@@ -22,7 +52,6 @@ class Flags(object):
     skip     = 1 << 5  # don't run example at all (e.g. notebooks are completely broken)
     no_js    = 1 << 6  # skip bokehjs and thus image diff (e.g. google maps key issue)
     no_diff  = 1 << 7  # skip only image diff (e.g. inherent randomness as in jitter)
-
 
 class Example(object):
 
@@ -252,3 +281,15 @@ def collect_examples(config_path):
             example_type=example_type, slow=slow_status, skip=skip_status, no_js=no_js_status, no_diff=no_diff_status)
 
     return list_of_examples
+
+#-----------------------------------------------------------------------------
+# Dev API
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Private API
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Code
+#-----------------------------------------------------------------------------
