+++ bokeh/testing/filesystem.py
@@ -34,6 +34,10 @@ import tempfile
 # Bokeh imports
 
 #-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
