+++ bokeh/testing/fixtures.py
@@ -29,6 +29,9 @@ import pytest
 # Bokeh imports
 from ..util.dependencies import import_optional
 
+#-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
 
 #-----------------------------------------------------------------------------
 # General API
