+++ bokeh/testing/git.py
@@ -31,6 +31,10 @@ import sys
 from bokeh.util.terminal import write
 
 #-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
