+++ bokeh/testing/images.py
@@ -5,7 +5,7 @@
 #
 # The full license is in the file LICENSE.txt, distributed with this software.
 #-----------------------------------------------------------------------------
-''' Provide utility classes and functions useful for testing Bokeh itself.
+''' Provide tools for testing Bokeh maipulating images.
 
 '''
 
@@ -22,11 +22,9 @@ log = logging.getLogger(__name__)
 #-----------------------------------------------------------------------------
 
 # Standard library imports
-import importlib
 
 # External imports
-import pytest
-from six import string_types
+from PIL import Image, ImageChops
 
 # Bokeh imports
 
@@ -38,28 +36,44 @@ from six import string_types
 # General API
 #-----------------------------------------------------------------------------
 
-def verify_all(module, ALL):
-    '''
+def image_diff(diff_path, before_path, after_path, superimpose=False):
+    """ Returns the percentage of differing pixels. """
+    before = Image.open(before_path)
+    after = Image.open(after_path)
 
-    '''
-    class Test___all__(object):
-        def test___all__(self):
-            if isinstance(module, string_types):
-                mod = importlib.import_module(module)
-            else:
-                mod = module
-            assert hasattr(mod, "__all__")
-            assert mod.__all__ == ALL
+    before = before.convert('RGBA')
+    after = after.convert('RGBA')
 
-        @pytest.mark.parametrize('name', ALL)
-        @pytest.mark.unit
-        def test_contents(self, name):
-            if isinstance(module, string_types):
-                mod = importlib.import_module(module)
-            else:
-                mod = module
-            assert hasattr(mod, name)
-    return Test___all__
+    width = max(before.width, after.width)
+    height = max(before.height, after.height)
+
+    resized_before = Image.new("RGBA", (width, height), "white")
+    resized_after = Image.new("RGBA", (width, height), "white")
+
+    resized_before.paste(before)
+    resized_after.paste(after)
+
+    mask = ImageChops.difference(resized_before, resized_after)
+    mask = mask.convert('L')
+    mask = mask.point(lambda k: 0 if k == 0 else 255)
+
+    if mask.getbbox() is None:
+        return 0
+    else:
+        diff = mask.convert('RGB')
+        if superimpose:
+            diff.paste(resized_after, mask=mask)
+        else:
+            diff.paste((0, 0, 255), mask=mask)
+        diff.save(diff_path)
+
+        pixels = 0
+
+        for v in mask.getdata():
+            if v == 255:
+                pixels += 1
+
+        return float(pixels)/(width*height)*100
 
 #-----------------------------------------------------------------------------
 # Dev API
@@ -72,3 +86,5 @@ def verify_all(module, ALL):
 #-----------------------------------------------------------------------------
 # Code
 #-----------------------------------------------------------------------------
+
+logging.getLogger('PIL.PngImagePlugin').setLevel(logging.INFO)
