+++ bokeh/testing/plugins/__init__.py
@@ -28,6 +28,10 @@ log = logging.getLogger(__name__)
 # Bokeh imports
 
 #-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
