+++ bokeh/testing/plugins/chrome_screenshot.js
