+++ bokeh/testing/plugins/examples_report.jinja
