+++ bokeh/testing/plugins/implicit_mark.py
@@ -29,6 +29,10 @@ from _pytest.mark import matchmark # TODO (bev) non-private API?
 # Bokeh imports
 
 #-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
