+++ bokeh/testing/plugins/screenshot.py
