+++ bokeh/testing/selenium.py
@@ -32,6 +32,10 @@ from selenium.webdriver.support.ui import WebDriverWait
 from bokeh.util.terminal import warn
 
 #-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
