+++ bokeh/testing/travis.py
@@ -29,6 +29,10 @@ import os
 # Bokeh imports
 
 #-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
 # General API
 #-----------------------------------------------------------------------------
